Network.AWS.SDB.BatchDeleteAttributes

Description

Performs multiple DeleteAttributes operations in a single call, which reduces round trips and latencies. This enables Amazon SimpleDB to optimize requests, which generally yields better throughput.

The following limitations are enforced for this operation: 1 MB request size 25 item limit per BatchDeleteAttributes operation

http://docs.aws.amazon.com/AmazonSimpleDB/latest/DeveloperGuide/SDB_API_BatchDeleteAttributes.html

Request

data BatchDeleteAttributes

Request constructor

batchDeleteAttributes

Arguments

:: Text
bdaDomainName

-> BatchDeleteAttributes

BatchDeleteAttributes constructor.

The fields accessible through corresponding lenses are:

bdaDomainName :: Text

bdaItems :: [DeletableItem]

Request lenses

bdaDomainName :: Lens' BatchDeleteAttributes Text

The name of the domain in which the attributes are being deleted.

bdaItems :: Lens' BatchDeleteAttributes [DeletableItem]

A list of items on which to perform the operation.

Response

data BatchDeleteAttributesResponse

Response constructor

batchDeleteAttributesResponse :: BatchDeleteAttributesResponse

BatchDeleteAttributesResponse constructor.