Network.AWS.SDB.CreateDomain

Description

The CreateDomain operation creates a new domain. The domain name should be unique among the domains associated with the Access Key ID provided in the request. The CreateDomain operation may take 10 or more seconds to complete.

The client can create up to 100 domains per account.

If the client requires additional domains, go to http://aws.amazon.com/contact-us/simpledb-limit-request/.

http://docs.aws.amazon.com/AmazonSimpleDB/latest/DeveloperGuide/SDB_API_CreateDomain.html

Request

data CreateDomain

Request constructor

createDomain

Arguments

:: Text
cdDomainName

-> CreateDomain

CreateDomain constructor.

The fields accessible through corresponding lenses are:

cdDomainName :: Text

Request lenses

cdDomainName :: Lens' CreateDomain Text

The name of the domain to create. The name can range between 3 and 255 characters and can contain the following characters: a-z, A-Z, 0-9, '_', -, and ..

Response

data CreateDomainResponse

Response constructor

createDomainResponse :: CreateDomainResponse

CreateDomainResponse constructor.