Network.AWS.SDB.DeleteAttributes

Description

Deletes one or more attributes associated with an item. If all attributes of the item are deleted, the item is deleted.

DeleteAttributes is an idempotent operation; running it multiple times on the same item or attribute does not result in an error response.

Because Amazon SimpleDB makes multiple copies of item data and uses an eventual consistency update model, performing a GetAttributes or Select operation (read) immediately after a DeleteAttributes or PutAttributes operation (write) might not return updated item data.

http://docs.aws.amazon.com/AmazonSimpleDB/latest/DeveloperGuide/SDB_API_DeleteAttributes.html

Request

data DeleteAttributes

Request constructor

deleteAttributes

Arguments

:: Text
daDomainName

-> Text
daItemName

-> DeleteAttributes

DeleteAttributes constructor.

The fields accessible through corresponding lenses are:

daAttributes :: [Attribute]

daDomainName :: Text

daExpected :: Maybe UpdateCondition

daItemName :: Text

Request lenses

daAttributes :: Lens' DeleteAttributes [Attribute]

A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

daDomainName :: Lens' DeleteAttributes Text

The name of the domain in which to perform the operation.

daExpected :: Lens' DeleteAttributes (Maybe UpdateCondition)

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

daItemName :: Lens' DeleteAttributes Text

The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

Response

data DeleteAttributesResponse

Response constructor

deleteAttributesResponse :: DeleteAttributesResponse

DeleteAttributesResponse constructor.