Network.AWS.SDB.DeleteDomain

Description

The DeleteDomain operation deletes a domain. Any items (and their attributes) in the domain are deleted as well. The DeleteDomain operation might take 10 or more seconds to complete.

Creating a Request

deleteDomain

Arguments

:: Text
ddDomainName

-> DeleteDomain

Creates a value of DeleteDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddDomainName - The name of the domain to delete.

data DeleteDomain

See: deleteDomain smart constructor.

Request Lenses

ddDomainName :: Lens' DeleteDomain Text

The name of the domain to delete.

Destructuring the Response

deleteDomainResponse :: DeleteDomainResponse

Creates a value of DeleteDomainResponse with the minimum fields required to make a request.

data DeleteDomainResponse

See: deleteDomainResponse smart constructor.