Network.AWS.SDB.DomainMetadata

Description

Returns information about the domain, including when the domain was created, the number of items and attributes in the domain, and the size of the attribute names and values.

http://docs.aws.amazon.com/AmazonSimpleDB/latest/DeveloperGuide/SDB_API_DomainMetadata.html

Request

data DomainMetadata

Request constructor

domainMetadata

Arguments

:: Text
dmDomainName

-> DomainMetadata

DomainMetadata constructor.

The fields accessible through corresponding lenses are:

dmDomainName :: Text

Request lenses

dmDomainName :: Lens' DomainMetadata Text

The name of the domain for which to display the metadata of.

Response

data DomainMetadataResponse

Response constructor

domainMetadataResponse :: DomainMetadataResponse

DomainMetadataResponse constructor.

The fields accessible through corresponding lenses are:

dmrAttributeNameCount :: Maybe Int

dmrAttributeNamesSizeBytes :: Maybe Integer

dmrAttributeValueCount :: Maybe Int

dmrAttributeValuesSizeBytes :: Maybe Integer

dmrItemCount :: Maybe Int

dmrItemNamesSizeBytes :: Maybe Integer

dmrTimestamp :: Maybe Int

Response lenses

dmrAttributeNameCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of unique attribute names in the domain.

dmrAttributeNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all unique attribute names in the domain, in bytes.

dmrAttributeValueCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of all attribute name/value pairs in the domain.

dmrAttributeValuesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all attribute values in the domain, in bytes.

dmrItemCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of all items in the domain.

dmrItemNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all item names in the domain, in bytes.

dmrTimestamp :: Lens' DomainMetadataResponse (Maybe Int)

The data and time when metadata was calculated, in Epoch (UNIX) seconds.