Network.AWS.SDB.GetAttributes

Description

Returns all of the attributes associated with the specified item. Optionally, the attributes returned can be limited to one or more attributes by specifying an attribute name parameter.

If the item does not exist on the replica that was accessed for this operation, an empty set is returned. The system does not return an error as it cannot guarantee the item does not exist on other replicas.

http://docs.aws.amazon.com/AmazonSimpleDB/latest/DeveloperGuide/SDB_API_GetAttributes.html

Request

data GetAttributes

Request constructor

getAttributes

Arguments

:: Text
gaDomainName

-> Text
gaItemName

-> GetAttributes

GetAttributes constructor.

The fields accessible through corresponding lenses are:

gaAttributeNames :: [Text]

gaConsistentRead :: Maybe Bool

gaDomainName :: Text

gaItemName :: Text

Request lenses

gaAttributeNames :: Lens' GetAttributes [Text]

The names of the attributes.

gaConsistentRead :: Lens' GetAttributes (Maybe Bool)

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

gaDomainName :: Lens' GetAttributes Text

The name of the domain in which to perform the operation.

gaItemName :: Lens' GetAttributes Text

The name of the item.

Response

data GetAttributesResponse

Response constructor

getAttributesResponse :: GetAttributesResponse

GetAttributesResponse constructor.

The fields accessible through corresponding lenses are:

garAttributes :: [Attribute]

Response lenses

garAttributes :: Lens' GetAttributesResponse [Attribute]

The list of attributes returned by the operation.