Network.AWS.SDB.ListDomains

Description

The ListDomains operation lists all domains associated with the Access Key ID. It returns domain names up to the limit set by <#MaxNumberOfDomains MaxNumberOfDomains>. A NextToken is returned if there are more than MaxNumberOfDomains domains. Calling ListDomains successive times with the NextToken provided by the operation returns up to MaxNumberOfDomains more domain names with each successive operation call.

http://docs.aws.amazon.com/AmazonSimpleDB/latest/DeveloperGuide/SDB_API_ListDomains.html

Request

data ListDomains

Request constructor

listDomains :: ListDomains

ListDomains constructor.

The fields accessible through corresponding lenses are:

ldMaxNumberOfDomains :: Maybe Int

ldNextToken :: Maybe Text

Request lenses

ldMaxNumberOfDomains :: Lens' ListDomains (Maybe Int)

The maximum number of domain names you want returned. The range is 1 to 100. The default setting is 100.

ldNextToken :: Lens' ListDomains (Maybe Text)

A string informing Amazon SimpleDB where to start the next list of domain names.

Response

data ListDomainsResponse

Response constructor

listDomainsResponse :: ListDomainsResponse

ListDomainsResponse constructor.

The fields accessible through corresponding lenses are:

ldrDomainNames :: [Text]

ldrNextToken :: Maybe Text

Response lenses

ldrDomainNames :: Lens' ListDomainsResponse [Text]

A list of domain names that match the expression.

ldrNextToken :: Lens' ListDomainsResponse (Maybe Text)

An opaque token indicating that there are more domains than the specified MaxNumberOfDomains still available.