Network.AWS.SDB.Select

Description

The Select operation returns a set of attributes for ItemNames that match the select expression. Select is similar to the standard SQL SELECT statement.

The total size of the response cannot exceed 1 MB in total size. Amazon SimpleDB automatically adjusts the number of items returned per page to enforce this limit. For example, if the client asks to retrieve 2500 items, but each individual item is 10 kB in size, the system returns 100 items and an appropriate NextToken so the client can access the next page of results.

For information on how to construct select expressions, see Using Select to Create Amazon SimpleDB Queries in the Developer Guide.

http://docs.aws.amazon.com/AmazonSimpleDB/latest/DeveloperGuide/SDB_API_Select.html

Request

data Select

Request constructor

select

Arguments

:: Text
sSelectExpression

-> Select

Select constructor.

The fields accessible through corresponding lenses are:

sConsistentRead :: Maybe Bool

sNextToken :: Maybe Text

sSelectExpression :: Text

Request lenses

sConsistentRead :: Lens' Select (Maybe Bool)

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

sNextToken :: Lens' Select (Maybe Text)

A string informing Amazon SimpleDB where to start the next list of ItemNames.

sSelectExpression :: Lens' Select Text

The expression used to query the domain.

Response

data SelectResponse

Response constructor

selectResponse :: SelectResponse

SelectResponse constructor.

The fields accessible through corresponding lenses are:

srItems :: [Item]

srNextToken :: Maybe Text

Response lenses

srItems :: Lens' SelectResponse [Item]

A list of items that match the select expression.

srNextToken :: Lens' SelectResponse (Maybe Text)

An opaque token indicating that more items than MaxNumberOfItems were matched, the response size exceeded 1 megabyte, or the execution time exceeded 5 seconds.