Network.AWS.SDB.Types

Service

data SDB

Version 2009-04-15 of the Amazon SimpleDB service.

Error

data RESTError :: *

XML

ns :: Text

Attribute

data Attribute

attribute

Arguments

:: Text
aName

-> Text
aValue

-> Attribute

Attribute constructor.

The fields accessible through corresponding lenses are:

aAlternateNameEncoding :: Maybe Text

aAlternateValueEncoding :: Maybe Text

aName :: Text

aValue :: Text

aAlternateNameEncoding :: Lens' Attribute (Maybe Text)

aAlternateValueEncoding :: Lens' Attribute (Maybe Text)

aName :: Lens' Attribute Text

The name of the attribute.

aValue :: Lens' Attribute Text

The value of the attribute.

DeletableItem

data DeletableItem

deletableItem

Arguments

:: Text
diName

-> DeletableItem

DeletableItem constructor.

The fields accessible through corresponding lenses are:

diAttributes :: [Attribute]

diName :: Text

diAttributes :: Lens' DeletableItem [Attribute]

diName :: Lens' DeletableItem Text

ReplaceableItem

data ReplaceableItem

replaceableItem

Arguments

:: Text
riName

-> ReplaceableItem

ReplaceableItem constructor.

The fields accessible through corresponding lenses are:

riAttributes :: [ReplaceableAttribute]

riName :: Text

riAttributes :: Lens' ReplaceableItem [ReplaceableAttribute]

The list of attributes for a replaceable item.

riName :: Lens' ReplaceableItem Text

The name of the replaceable item.

UpdateCondition

data UpdateCondition

updateCondition :: UpdateCondition

UpdateCondition constructor.

The fields accessible through corresponding lenses are:

ucExists :: Maybe Bool

ucName :: Maybe Text

ucValue :: Maybe Text

ucExists :: Lens' UpdateCondition (Maybe Bool)

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

ucName :: Lens' UpdateCondition (Maybe Text)

The name of the attribute involved in the condition.

ucValue :: Lens' UpdateCondition (Maybe Text)

The value of an attribute. This value can only be specified when the Exists parameter is equal to true.

ReplaceableAttribute

data ReplaceableAttribute

replaceableAttribute

Arguments

:: Text
raName

-> Text
raValue

-> ReplaceableAttribute

ReplaceableAttribute constructor.

The fields accessible through corresponding lenses are:

raName :: Text

raReplace :: Maybe Bool

raValue :: Text

raName :: Lens' ReplaceableAttribute Text

The name of the replaceable attribute.

raReplace :: Lens' ReplaceableAttribute (Maybe Bool)

A flag specifying whether or not to replace the attribute/value pair or to add a new attribute/value pair. The default setting is false.

raValue :: Lens' ReplaceableAttribute Text

The value of the replaceable attribute.

Item

data Item

item

Arguments

:: Text
iName

-> Item

Item constructor.

The fields accessible through corresponding lenses are:

iAlternateNameEncoding :: Maybe Text

iAttributes :: [Attribute]

iName :: Text

iAlternateNameEncoding :: Lens' Item (Maybe Text)

iAttributes :: Lens' Item [Attribute]

A list of attributes.

iName :: Lens' Item Text

The name of the item.