Network.AWS.SDB.BatchDeleteAttributes

Description

Performs multiple DeleteAttributes operations in a single call, which reduces round trips and latencies. This enables Amazon SimpleDB to optimize requests, which generally yields better throughput.

The following limitations are enforced for this operation: * 1 MB request size * 25 item limit per BatchDeleteAttributes operation

Creating a Request

batchDeleteAttributes

Arguments

:: Text
bdaDomainName

-> BatchDeleteAttributes

Creates a value of BatchDeleteAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bdaDomainName - The name of the domain in which the attributes are being deleted.

bdaItems - A list of items on which to perform the operation.

data BatchDeleteAttributes

See: batchDeleteAttributes smart constructor.

Request Lenses

bdaDomainName :: Lens' BatchDeleteAttributes Text

The name of the domain in which the attributes are being deleted.

bdaItems :: Lens' BatchDeleteAttributes [DeletableItem]

A list of items on which to perform the operation.

Destructuring the Response

batchDeleteAttributesResponse :: BatchDeleteAttributesResponse

Creates a value of BatchDeleteAttributesResponse with the minimum fields required to make a request.

data BatchDeleteAttributesResponse

See: batchDeleteAttributesResponse smart constructor.