Network.AWS.SDB.DeleteAttributes

Description

Deletes one or more attributes associated with an item. If all attributes of the item are deleted, the item is deleted.

DeleteAttributes is an idempotent operation; running it multiple times on the same item or attribute does not result in an error response.

Because Amazon SimpleDB makes multiple copies of item data and uses an eventual consistency update model, performing a GetAttributes or Select operation (read) immediately after a DeleteAttributes or PutAttributes operation (write) might not return updated item data.

Creating a Request

deleteAttributes

Arguments

:: Text
daDomainName

-> Text
daItemName

-> DeleteAttributes

Creates a value of DeleteAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daAttributes - A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

daExpected - The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

daDomainName - The name of the domain in which to perform the operation.

daItemName - The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

data DeleteAttributes

See: deleteAttributes smart constructor.

Request Lenses

daAttributes :: Lens' DeleteAttributes [Attribute]

A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

daExpected :: Lens' DeleteAttributes (Maybe UpdateCondition)

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

daDomainName :: Lens' DeleteAttributes Text

The name of the domain in which to perform the operation.

daItemName :: Lens' DeleteAttributes Text

The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

Destructuring the Response

deleteAttributesResponse :: DeleteAttributesResponse

Creates a value of DeleteAttributesResponse with the minimum fields required to make a request.

data DeleteAttributesResponse

See: deleteAttributesResponse smart constructor.