Network.AWS.SDB.DomainMetadata

Description

Returns information about the domain, including when the domain was created, the number of items and attributes in the domain, and the size of the attribute names and values.

Creating a Request

domainMetadata

Arguments

:: Text
dmDomainName

-> DomainMetadata

Creates a value of DomainMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmDomainName - The name of the domain for which to display the metadata of.

data DomainMetadata

See: domainMetadata smart constructor.

Request Lenses

dmDomainName :: Lens' DomainMetadata Text

The name of the domain for which to display the metadata of.

Destructuring the Response

domainMetadataResponse

Arguments

:: Int
dmrsResponseStatus

-> DomainMetadataResponse

Creates a value of DomainMetadataResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmrsItemNamesSizeBytes - The total size of all item names in the domain, in bytes.

dmrsAttributeValuesSizeBytes - The total size of all attribute values in the domain, in bytes.

dmrsAttributeNameCount - The number of unique attribute names in the domain.

dmrsAttributeNamesSizeBytes - The total size of all unique attribute names in the domain, in bytes.

dmrsAttributeValueCount - The number of all attribute name/value pairs in the domain.

dmrsItemCount - The number of all items in the domain.

dmrsTimestamp - The data and time when metadata was calculated, in Epoch (UNIX) seconds.

dmrsResponseStatus - -- | The response status code.

data DomainMetadataResponse

See: domainMetadataResponse smart constructor.

Response Lenses

dmrsItemNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all item names in the domain, in bytes.

dmrsAttributeValuesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all attribute values in the domain, in bytes.

dmrsAttributeNameCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of unique attribute names in the domain.

dmrsAttributeNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all unique attribute names in the domain, in bytes.

dmrsAttributeValueCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of all attribute name/value pairs in the domain.

dmrsItemCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of all items in the domain.

dmrsTimestamp :: Lens' DomainMetadataResponse (Maybe Int)

The data and time when metadata was calculated, in Epoch (UNIX) seconds.

dmrsResponseStatus :: Lens' DomainMetadataResponse Int

- | The response status code.