Network.AWS.SDB.GetAttributes

Description

Returns all of the attributes associated with the specified item. Optionally, the attributes returned can be limited to one or more attributes by specifying an attribute name parameter.

If the item does not exist on the replica that was accessed for this operation, an empty set is returned. The system does not return an error as it cannot guarantee the item does not exist on other replicas.

Creating a Request

getAttributes

Arguments

:: Text
gaDomainName

-> Text
gaItemName

-> GetAttributes

Creates a value of GetAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaConsistentRead - true

gaAttributeNames - The names of the attributes.

gaDomainName - The name of the domain in which to perform the operation.

gaItemName - The name of the item.

data GetAttributes

See: getAttributes smart constructor.

Request Lenses

gaConsistentRead :: Lens' GetAttributes (Maybe Bool)

true

gaAttributeNames :: Lens' GetAttributes [Text]

The names of the attributes.

gaDomainName :: Lens' GetAttributes Text

The name of the domain in which to perform the operation.

gaItemName :: Lens' GetAttributes Text

The name of the item.

Destructuring the Response

getAttributesResponse

Arguments

:: Int
garsResponseStatus

-> GetAttributesResponse

Creates a value of GetAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

garsAttributes - The list of attributes returned by the operation.

garsResponseStatus - -- | The response status code.

data GetAttributesResponse

See: getAttributesResponse smart constructor.

Response Lenses

garsAttributes :: Lens' GetAttributesResponse [Attribute]

The list of attributes returned by the operation.

garsResponseStatus :: Lens' GetAttributesResponse Int

- | The response status code.