Network.AWS.SDB.Select

Description

The Select operation returns a set of attributes for ItemNames that match the select expression. Select is similar to the standard SQL SELECT statement.

The total size of the response cannot exceed 1 MB in total size. Amazon SimpleDB automatically adjusts the number of items returned per page to enforce this limit. For example, if the client asks to retrieve 2500 items, but each individual item is 10 kB in size, the system returns 100 items and an appropriate NextToken so the client can access the next page of results.

For information on how to construct select expressions, see Using Select to Create Amazon SimpleDB Queries in the Developer Guide.

This operation returns paginated results.

Creating a Request

select

Arguments

:: Text
sSelectExpression

-> Select

Creates a value of Select with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sConsistentRead - true

sNextToken - ItemNames

sSelectExpression - The expression used to query the domain.

data Select

See: select smart constructor.

Request Lenses

sConsistentRead :: Lens' Select (Maybe Bool)

true

sNextToken :: Lens' Select (Maybe Text)

ItemNames

sSelectExpression :: Lens' Select Text

The expression used to query the domain.

Destructuring the Response

selectResponse

Arguments

:: Int
srsResponseStatus

-> SelectResponse

Creates a value of SelectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srsItems - A list of items that match the select expression.

srsNextToken - MaxNumberOfItems

srsResponseStatus - -- | The response status code.

data SelectResponse

See: selectResponse smart constructor.

Response Lenses

srsItems :: Lens' SelectResponse [Item]

A list of items that match the select expression.

srsNextToken :: Lens' SelectResponse (Maybe Text)

MaxNumberOfItems

srsResponseStatus :: Lens' SelectResponse Int

- | The response status code.