Network.AWS.SDB.Types

Description

Service Configuration

sdb :: Service

API version 2009-04-15 of the Amazon SimpleDB SDK configuration.

Errors

_InvalidNumberValueTests :: AsError a => Getting (First ServiceError) a ServiceError

Too many predicates exist in the query expression.

_NoSuchDomain :: AsError a => Getting (First ServiceError) a ServiceError

The specified domain does not exist.

_NumberSubmittedItemsExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many items exist in a single call.

_AttributeDoesNotExist :: AsError a => Getting (First ServiceError) a ServiceError

The specified attribute does not exist.

_NumberDomainAttributesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes in this domain.

_DuplicateItemName :: AsError a => Getting (First ServiceError) a ServiceError

The item name was specified more than once.

_MissingParameter :: AsError a => Getting (First ServiceError) a ServiceError

The request must contain the specified missing parameter.

_InvalidNextToken :: AsError a => Getting (First ServiceError) a ServiceError

The specified NextToken is not valid.

_InvalidParameterValue :: AsError a => Getting (First ServiceError) a ServiceError

The value for a parameter is invalid.

_NumberItemAttributesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes in this item.

_RequestTimeout :: AsError a => Getting (First ServiceError) a ServiceError

A timeout occurred when attempting to query the specified domain with specified query expression.

_TooManyRequestedAttributes :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes requested.

_InvalidNumberPredicates :: AsError a => Getting (First ServiceError) a ServiceError

Too many predicates exist in the query expression.

_NumberDomainsExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many domains exist per this account.

_NumberSubmittedAttributesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes exist in a single call.

_NumberDomainBytesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many bytes in this domain.

_InvalidQueryExpression :: AsError a => Getting (First ServiceError) a ServiceError

The specified query expression syntax is not valid.

Attribute

data Attribute

See: attribute smart constructor.

attribute

Arguments

:: Text
aName

-> Text
aValue

-> Attribute

Creates a value of Attribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aAlternateValueEncoding -

aAlternateNameEncoding -

aName - The name of the attribute.

aValue - The value of the attribute.

aAlternateValueEncoding :: Lens' Attribute (Maybe Text)

aAlternateNameEncoding :: Lens' Attribute (Maybe Text)

aName :: Lens' Attribute Text

The name of the attribute.

aValue :: Lens' Attribute Text

The value of the attribute.

DeletableItem

data DeletableItem

See: deletableItem smart constructor.

deletableItem

Arguments

:: Text
diName

-> DeletableItem

Creates a value of DeletableItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diAttributes - Undocumented member.

diName - Undocumented member.

diAttributes :: Lens' DeletableItem [Attribute]

Undocumented member.

diName :: Lens' DeletableItem Text

Undocumented member.

Item

data Item

See: item smart constructor.

item

Arguments

:: Text
iName

-> Item

Creates a value of Item with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iAlternateNameEncoding -

iName - The name of the item.

iAttributes - A list of attributes.

iAlternateNameEncoding :: Lens' Item (Maybe Text)

iName :: Lens' Item Text

The name of the item.

iAttributes :: Lens' Item [Attribute]

A list of attributes.

ReplaceableAttribute

data ReplaceableAttribute

See: replaceableAttribute smart constructor.

replaceableAttribute

Arguments

:: Text
raName

-> Text
raValue

-> ReplaceableAttribute

Creates a value of ReplaceableAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

raReplace - false

raName - The name of the replaceable attribute.

raValue - The value of the replaceable attribute.

raReplace :: Lens' ReplaceableAttribute (Maybe Bool)

false

raName :: Lens' ReplaceableAttribute Text

The name of the replaceable attribute.

raValue :: Lens' ReplaceableAttribute Text

The value of the replaceable attribute.

ReplaceableItem

data ReplaceableItem

See: replaceableItem smart constructor.

replaceableItem

Arguments

:: Text
riName

-> ReplaceableItem

Creates a value of ReplaceableItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riName - The name of the replaceable item.

riAttributes - The list of attributes for a replaceable item.

riName :: Lens' ReplaceableItem Text

The name of the replaceable item.

riAttributes :: Lens' ReplaceableItem [ReplaceableAttribute]

The list of attributes for a replaceable item.

UpdateCondition

data UpdateCondition

Specifies the conditions under which data should be updated. If an update condition is specified for a request, the data will only be updated if the condition is satisfied. For example, if an attribute with a specific name and value exists, or if a specific attribute doesn't exist.

See: updateCondition smart constructor.

updateCondition :: UpdateCondition

Creates a value of UpdateCondition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucExists - A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

ucValue - The value of an attribute. This value can only be specified when the Exists parameter is equal to true .

ucName - The name of the attribute involved in the condition.

ucExists :: Lens' UpdateCondition (Maybe Bool)

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

ucValue :: Lens' UpdateCondition (Maybe Text)

The value of an attribute. This value can only be specified when the Exists parameter is equal to true .

ucName :: Lens' UpdateCondition (Maybe Text)

The name of the attribute involved in the condition.